backend/test/requirements.txt
binance-futures-connector==4.1.0
python-dotenv==1.0.1